feat(speakers): redesign speaker drawer with tabbed layout and sub-views (Story 10.30)
Replace 3 monolithic drawers (1,713 lines) with a single tabbed SpeakerDetailDrawer
containing Overview/Details/Activity tabs, inline organizer assignment, and in-drawer
sub-views for content submission and quality review. Improves mobile UX with full-width
drawer, scrollable tabs, and sticky action footer.
