+++ _bmad-output/implementation-artifacts/sprint-status.yaml
@@ -170,6 +170,7 @@ development_status:
   10-26-ses-email-forwarding-distribution-lists: done                          # SES+Lambda forwarding for ok@/info@/events@/partner@/support@/batbern{N}@batbern.ch; Admin Settings tab; sender auth; prereq: 10.17
   10-27-ical-rsvp-tracking-partner-meetings: done                              # iCal REPLY ingestion; partner_meeting_rsvps table (V9); POST /internal/partner-meetings/rsvps; GET /api/v1/partner-meetings/{id}/rsvps; PartnerMeetingRsvpPanel; prereq: 8.3, 10.17
   10-28-newsletter-subscriber-management: done                                 # organizer nav item; paginated/searchable/sortable subscriber list; unsubscribe/resubscribe/delete actions; backend sort; tech debt: UserTable backend sort; prereq: 10.7
+  10-30-speaker-drawer-redesign: review                                   # tabbed drawer (Overview/Details/Activity); inline organizer assignment; sub-views for content submission + quality review; mobile UX; replaces 3 monolithic drawers
   epic-10-retrospective: optional
 
   # ============================================================
