+++ web-frontend/public/locales/fi/organizer.json
@@ -324,7 +324,22 @@
     "sendReminder": "Lähetä muistutus",
     "sendingReminder": "Lähetetään...",
     "reminderSent": "Muistutus lähetetty osoitteeseen {{email}}",
-    "reminderFailed": "Muistutuksen lähettäminen epäonnistui"
+    "reminderFailed": "Muistutuksen lähettäminen epäonnistui",
+    "tabs": {
+      "overview": "Overview",
+      "details": "Details",
+      "activity": "Activity"
+    },
+    "drawer": {
+      "back": "Back"
+    },
+    "organizerUpdated": "Organizer updated",
+    "organizerUpdateFailed": "Failed to update organizer",
+    "submitContent": "Submit Content",
+    "reviewContent": "Review Content",
+    "revisionRequested": "Revision Requested",
+    "submittedContent": "Submitted Content",
+    "submittedAt": "Submitted"
   },
   "tasks": {
     "taskBoard": "Tehtävätaulu",
