+++ web-frontend/public/locales/fr/organizer.json
@@ -324,7 +324,22 @@
     "sendReminder": "Envoyer un rappel",
     "sendingReminder": "Envoi en cours...",
     "reminderSent": "Rappel envoyé à {{email}}",
-    "reminderFailed": "Échec de l'envoi du rappel"
+    "reminderFailed": "Échec de l'envoi du rappel",
+    "tabs": {
+      "overview": "Overview",
+      "details": "Details",
+      "activity": "Activity"
+    },
+    "drawer": {
+      "back": "Back"
+    },
+    "organizerUpdated": "Organizer updated",
+    "organizerUpdateFailed": "Failed to update organizer",
+    "submitContent": "Submit Content",
+    "reviewContent": "Review Content",
+    "revisionRequested": "Revision Requested",
+    "submittedContent": "Submitted Content",
+    "submittedAt": "Submitted"
   },
   "tasks": {
     "taskBoard": "Tableau des tâches",
