+++ web-frontend/src/components/organizer/EventPage/EventSpeakersTab.tsx
@@ -45,9 +45,7 @@ import { useQueryClient } from '@tanstack/react-query';
 import { SpeakerStatusLanes } from '@/components/organizer/SpeakerStatus/SpeakerStatusLanes';
 import { SpeakersSessionsTable } from '@/components/organizer/EventManagement/SpeakersSessionsTable';
 import { SpeakerBrainstormingPanel } from '@/components/SpeakerBrainstormingPanel/SpeakerBrainstormingPanel';
-import SpeakerOutreachDetailsDrawer from '@/components/organizer/SpeakerOutreach/SpeakerOutreachDetailsDrawer';
-import { ContentSubmissionDrawer } from '@/components/organizer/SpeakerStatus/ContentSubmissionDrawer';
-import { QualityReviewDrawer } from '@/components/organizer/SpeakerStatus/QualityReviewDrawer';
+import { SpeakerDetailDrawer } from '@/components/organizer/SpeakerDrawer';
 import type { SpeakerPoolEntry } from '@/types/speakerPool.types';
 import type { SessionUI, SessionSpeaker } from '@/types/event.types';
 import type { SessionUpdateData } from '@/components/organizer/EventManagement/SessionEditModal';
@@ -74,12 +72,6 @@ export const EventSpeakersTab: React.FC<EventSpeakersTabProps> = ({ eventCode })
   const [detailsDrawerOpen, setDetailsDrawerOpen] = useState(false);
   const [autoAssignLoading, setAutoAssignLoading] = useState(false);
   const [autoAssignError, setAutoAssignError] = useState<string | null>(null);
-
-  // Content submission and quality review drawer state
-  const [contentDrawerOpen, setContentDrawerOpen] = useState(false);
-  const [contentSpeaker, setContentSpeaker] = useState<SpeakerPoolEntry | null>(null);
-  const [reviewDrawerOpen, setReviewDrawerOpen] = useState(false);
-  const [reviewSpeaker, setReviewSpeaker] = useState<SpeakerPoolEntry | null>(null);
 
   // Fetch speaker status summary
   const { data: summary, isLoading: summaryLoading } = useQuery({
@@ -144,18 +136,6 @@ export const EventSpeakersTab: React.FC<EventSpeakersTabProps> = ({ eventCode })
   const handleIdentifiedToContacted = (speaker: SpeakerPoolEntry) => {
     setSelectedSpeaker(speaker);
     setDetailsDrawerOpen(true);
-  };
-
-  // Handle opening content submission drawer (from outreach details drawer)
-  const handleOpenContentSubmission = (speaker: SpeakerPoolEntry) => {
-    setContentSpeaker(speaker);
-    setContentDrawerOpen(true);
-  };
-
-  // Handle opening quality review drawer (from outreach details drawer)
-  const handleOpenQualityReview = (speaker: SpeakerPoolEntry) => {
-    setReviewSpeaker(speaker);
-    setReviewDrawerOpen(true);
   };
 
   // Session handlers
@@ -405,32 +385,11 @@ export const EventSpeakersTab: React.FC<EventSpeakersTabProps> = ({ eventCode })
         </Paper>
       )}
 
-      {/* Speaker Details Drawer (with contact history for all statuses) */}
-      <SpeakerOutreachDetailsDrawer
+      {/* Speaker Detail Drawer (tabbed: Overview, Details, Activity + sub-views) */}
+      <SpeakerDetailDrawer
         open={detailsDrawerOpen}
         onClose={() => setDetailsDrawerOpen(false)}
         speaker={selectedSpeaker}
-        eventCode={eventCode}
-        showMarkContactedForm={
-          selectedSpeaker?.status === 'IDENTIFIED' || selectedSpeaker?.status === 'CONTACTED'
-        }
-        onOpenContentSubmission={handleOpenContentSubmission}
-        onOpenQualityReview={handleOpenQualityReview}
-      />
-
-      {/* Content Submission Drawer (for ACCEPTED speakers) */}
-      <ContentSubmissionDrawer
-        open={contentDrawerOpen}
-        onClose={() => setContentDrawerOpen(false)}
-        speaker={contentSpeaker}
-        eventCode={eventCode}
-      />
-
-      {/* Quality Review Drawer (for CONTENT_SUBMITTED speakers) */}
-      <QualityReviewDrawer
-        open={reviewDrawerOpen}
-        onClose={() => setReviewDrawerOpen(false)}
-        speaker={reviewSpeaker}
         eventCode={eventCode}
       />
     </Stack>
