+++ web-frontend/src/components/organizer/SpeakerDrawer/index.ts
@@ -1,0 +1,1 @@
+export { SpeakerDetailDrawer } from './SpeakerDetailDrawer';
